-Wsign-compare warnings fix, the warning is disabled

+++ projects/miopen/src/solver/batchnorm/backward_spatial.cpp
@@ -111,7 +111,7 @@ bool PerformanceConfigBnBwdBackward::SetNextValue(
         return true;
     }
     // Get next valid configuration
-    if((this->index + 1) < valid_kernels.size())
+    if(static_cast<size_t>(this->index + 1) < valid_kernels.size())
     {
         ++this->index;
         this->kernel_id = this->valid_kernels[index];
@@ -130,7 +130,7 @@ bool PerformanceConfigBnBwdBackward::operator==(const PerformanceConfigBnBwdBack
 
 bool PerformanceConfigBnBwdBackward::IsValidValue() const
 {
-    return this->index >= 0 && this->index < valid_kernels.size();
+    return this->index >= 0 && static_cast<size_t>(this->index) < valid_kernels.size();
 }
 
 bool BnBwdTrainingSpatial::IsApplicable(
